Language Modeling

Language Models

NGrams.LanguageModel — Type.

LanguageModel(N; bos, eos, estimator=NGrams.MLE())

Create an N-gram language model, estimating probabilities with estimator.

source

Training

NGrams.fit! — Function.

NGrams.fit!(lm::LanguageModel, tokens)

Train the language model by observing a sequence of tokens.

source

Probability and Smoothing

NGrams.MLE — Type.

NGrams.MLE()

Maximum Likelihood Estimation for n-gram language modeling.

source

NGrams.AddK — Type.

NGrams.AddK(k::Number)

Add-k probability smoothing for n-gram language modeling.

source

NGrams.Laplace — Type.

NGrams.Laplace()

Laplace (add-1) smoothing for n-gram language modeling.

source

NGrams.LinearInterpolation — Type.

LinearInterpolation(λ)

Linear interpolation for probability smoothing in n-gram language modeling.

source

NGrams.AbsoluteDiscounting — Type.

NGrams.AbsoluteDiscounting(d::Number)

Absolute discounting for n-gram language modeling.

source

Sampling from a language model

NGrams.sample — Function.

NGrams.sample([rng::AbstractRNG,] lm, [vocabulary])

Sample a single token from the language model.

source

NGrams.generate — Function.

NGrams.generate(lm, num_words=1, text_seed=[])

Randomly generate num_words from language model.

If text_seed is provided, output is conditioned on that history. The seed is included in the return value and counts against num_words.

source